+++ src/VaultSharp/V1/SecretsEngines/PKI/IPKISecretsEngine.cs
@@ -90,6 +90,20 @@ namespace VaultSharp.V1.SecretsEngines.PKI
         Task TidyAsync(CertificateTidyRequest certificateTidyRequest = null, string pkiBackendMountPoint = null);
 
         /// <summary>
+        /// This endpoint allows auto tidying up the storage backend and/or CRL by removing certificates that have expired 
+        /// and are past a certain buffer period beyond their expiration time.
+        /// </summary>
+        /// <param name="certificateAutoTidyRequest">The request object</param>
+        /// <param name="pkiBackendMountPoint"><para>[optional]</para>
+        /// The mount point for the PKI backend. Defaults to <see cref="SecretsEngineMountPoints.PKI" />
+        /// Provide a value only if you have customized the PKI mount point.
+        /// </param>
+        /// <returns>
+        /// The task
+        /// </returns>
+        Task AutoTidyAsync(CertificateAutoTidyRequest certificateAutoTidyRequest = null, string pkiBackendMountPoint = null);
+
+        /// <summary>
         /// Retrieves the CA certificate in raw DER-encoded form. 
         /// This is a bare endpoint that does not return a standard Vault data structure. 
         /// The CA certificate can be returned in DER or PEM format.