+++ src/VaultSharp/V1/SecretsEngines/PKI/PKISecretsEngineProvider.cs
@@ -51,6 +51,13 @@ namespace VaultSharp.V1.SecretsEngines.PKI
             await _polymath.MakeVaultApiRequest(pkiBackendMountPoint ?? _polymath.VaultClientSettings.SecretsEngineMountPoints.PKI, "/tidy", HttpMethod.Post, newRequest).ConfigureAwait(_polymath.VaultClientSettings.ContinueAsyncTasksOnCapturedContext);
         }
 
+        public async Task AutoTidyAsync(CertificateAutoTidyRequest certificateAutoTidyRequest = null, string pkiBackendMountPoint = null)
+        {
+            var newRequest = certificateAutoTidyRequest ?? new CertificateAutoTidyRequest();
+
+            await _polymath.MakeVaultApiRequest(pkiBackendMountPoint ?? _polymath.VaultClientSettings.SecretsEngineMountPoints.PKI, "/config/auto-tidy", HttpMethod.Post, newRequest).ConfigureAwait(_polymath.VaultClientSettings.ContinueAsyncTasksOnCapturedContext);
+        }
+
         public async Task<RawCertificateData> ReadCACertificateAsync(CertificateFormat certificateFormat = CertificateFormat.der, string pkiBackendMountPoint = null)
         {
             var format = certificateFormat == CertificateFormat.pem ? "/" + CertificateFormat.pem : string.Empty;