+++ CHANGELOG.md
@@ -7,12 +7,14 @@
 **FEATURES:**
 
   * secret/pki: Add new api to list revoked certificates
+  * secret/pki: Add new api to auto tidy certificates
 
 **IMPROVEMENTS:**
 
   * [GH-292](https://github.com/rajanadar/VaultSharp/issues/292) Add warning and other fields to seal status apis.
   * [GH-291](https://github.com/rajanadar/VaultSharp/issues/291) Add logger endpoints to System backend.
   * Added several fields like plugin_version etc. to the Backend fetching System APIs.
+  * Added several new fields to the PKI Cert Tidy Request payload.
 
 ## 1.12.2 (December 09, 2022)